Ordering tickets
1. OrderController and OrderSummartController have been created
for ordering tickets.
2. OrderRequest has been added for validation of order data.
3. OrderService has been created. It handles the order business logic.
4. In 'discounts' table, column 'value' is float-type now.
5. DiscountSeeder has been modified for testing data.
6. Views for ordering have been added.
7. JS files for the project have been configured in 'webpack.mid.js'.
8. File 'main.js' has been added to public. It contains data from
all JS files in the project (after compilation to dev/prod).
9. More translations have been added.

Project is ready to implement paypal payments.

 On branch reservations

+++ database/migrations/2020_10_20_213005_create_discounts_table.php
@@ -13,7 +13,7 @@ class CreateDiscountsTable extends Migration
             $table->bigInteger('user_id')->unsigned();
             $table->timestamps();
             $table->string('code', 16);
-            $table->tinyInteger('value');
+            $table->float('value', 3, 2);
         });
         
             Schema::table('discounts', function (Blueprint $table) {
